+++ packages/create-theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sugarat/create-theme
 
+## 0.0.107
+
+### Patch Changes
+
+- chore: update template theme version to `@sugarat/theme@^0.5.28`
+
 ## 0.0.106
 
 ### Patch Changes